backend/pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "FastAPI Boilerplate"
authors = ["Abbas <me@abbasmj.com>"]

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.9.4"
SQLAlchemy = "^2.0.4"
PyJWT = "^2.6.0"
uvicorn = "^0.20.0"
fastapi = "^0.104.1"
celery = "^5.2.7"
gunicorn = "^20.1.0"
fastapi-event = "^0.1.3"
ujson = "^5.7.0"
click = "^8.1.3"
greenlet = "1.1.3"
redis = "^4.5.4"
isort = "^5.12.0"
black = "^23.1.0"
pylint = "^2.16.2"
asyncpg = "^0.27.0"
pytest = "^7.2.1"
httpx = "^0.23.3"
pyyaml = "^6.0"
email-validator = "^1.3.1"
python-jose = "^3.3.0"
python-multipart = "0.0.6"
cryptography = "^41.0.0"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
faker = "^17.0.0"
pytest-asyncio = "^0.20.3"
aerich = "^0.7.1"
tortoise-orm="^0.20.0"
pydantic="^2.4.2"
pydantic-settings="^2.0.3"

[tool.poetry.dev-dependencies]
behave = "^1.2.6"

[tool.poetry.group.dev.dependencies]
black = {version = "^23.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
